syntax

+++ classify_change.py
@@ -14,7 +14,7 @@ def load_data():
 
 if __name__ == '__main__':
     X, labels = load_data()
-    x = np.linspace(0, 39750, 250, dtype=int)
+    x = np.linspace(0, 39750, 250, dtype=np.int16)
     
     x = np.concatenate(x, np.where(labels==1)[0])
     x = np.sort(np.array(x))
